Refactor seminarList component
